Move all related validation to validate.go
